chip8/requirements.txt
imgui[glfw]==2.0.0
PyOpenGL==3.1.6
numpy